route refactored into folder
