+++ server.js
@@ -4,7 +4,8 @@ const cors = require('cors');
 const knex = require('./knex');
 
 const { PORT, CLIENT_ORIGIN } = require('./config');
-// const { dbConnect } = require('./db-knex');
+
+const deckRouter = require('./routes/decks');
 
 // Create an Express application
 const app = express();
@@ -21,51 +22,7 @@ app.use(cors({ origin: CLIENT_ORIGIN }));
 // Parse request body
 app.use(express.json());
 
-//create route handler
-app.get('/api/cards/:uniqueUrl', function(req, res, next) {
-	console.log('backend');
-	const unique_url = req.params.uniqueUrl;
-	knex
-		.first('mtg_cards_id')
-		.from('cards')
-		.where('unique_url', unique_url)
-		.then(result => res.json(result))
-		.catch(err => next(err));
-});
-
-app.get('/api/cards', function(req, res, next) {
-	knex('cards')
-		.select('mtg_cards_id', 'unique_url')
-		.then(results => {
-			res.json(results);
-		})
-		.catch(err => next(err));
-});
-app.post('/api/cards', function(req, res, next) {
-	const { mtg_cards_id, unique_url } = req.body;
-	// console.info(req.body);
-	const newCard = {
-		mtg_cards_id: mtg_cards_id,
-		unique_url: unique_url,
-		decks_id: 1
-	};
-	// console.info(newCard);
-	if (newCard.mtg_cards_id === '[]') {
-		const err = new Error('Please add a card');
-		err.status = 400;
-		return next(err);
-	}
-	knex('cards')
-		.insert(newCard)
-		.into('cards')
-		.returning(['unique_url'])
-		.then(card => {
-			res.json(card[0]);
-		})
-		.catch(err => {
-			next(err);
-		});
-});
+app.use('/api/cards', deckRouter);
 
 // Custom 404 Not Found route handler
 app.use((req, res, next) => {
@@ -75,7 +32,7 @@ app.use((req, res, next) => {
 });
 
 // Custom Error Handler
-app.use((err, req, res, next) => {
+app.use((err, req, res) => {
 	if (err.status) {
 		const errBody = Object.assign({}, err, { message: err.message });
 		res.status(err.status).json(errBody);
